+++ asdf/extension.py
@@ -8,6 +8,7 @@ import warnings
 from pkg_resources import iter_entry_points
 
 import six
+import warnings
 import importlib
 
 from . import asdftypes
@@ -117,8 +118,12 @@ class AsdfExtensionList(object):
         for extension in extensions:
             if not isinstance(extension, AsdfExtension):
                 raise TypeError(
-                    "Extension must implement asdftypes.AsdfExtension "
-                    "interface")
+                    "Extension '{extension!r}' must implement the asdf.extension.AsdfExtension "
+                    "interface".format(extension=extension))
+            elif not issubclass(type(extension), AsdfExtension):
+                warnings.warn("Extensions not inheriting from asdf.extension.AsdfExtension"
+                              " are deprecated.",
+                              DeprecationWarning)
             tag_mapping.extend(extension.tag_mapping)
             url_mapping.extend(extension.url_mapping)
             for typ in extension.types: